feat: relocate nested dashboard sub-pages
