+++ apps/web/src/app/(authedLayout)/dashboard/layout.tsx
@@ -1,0 +1,5 @@
+import type { PropsWithChildren } from "react";
+
+export default function PageLayout({ children }: PropsWithChildren) {
+  return <section>{children}</section>;
+}
