+++ apps/web/src/app/(authedLayout)/dashboard/page.tsx
@@ -1,0 +1,7 @@
+export default function DashboardPage() {
+  return (
+    <article className="p-4 flex-1 h-full overflow-y-auto">
+      <p className="text-foreground">Dashboard page</p>
+    </article>
+  );
+}
