+++ apps/web/src/app/(authedLayout)/inbox/components/EmailEmptyView.tsx
