+++ apps/web/src/app/(authedLayout)/inbox/components/FolderColumn.tsx
@@ -8,23 +8,22 @@ type Props = { activeFolder: string };
 export const FolderColumn = ({ activeFolder }: Props) => {
   const router = useRouter();
   const context = UseInboxContext();
-  const folderList = Object.values(context!.folderMap);
 
   return (
     <div className="w-[280px] border-r border-gray-200 dark:border-gray-800 overflow-y-auto p-2 space-y-2">
       <ul className="space-y-1">
-        {folderList.map((folder) => {
+        {Object.values(context!.folderMap).map((folder) => {
           const path = encodeURIComponent(folder.name.toLowerCase());
           const folderName = folder.name.toLowerCase();
-          const isActive = activeFolder.toLowerCase() === folderName;
+
           return (
             <li
               key={folderName}
               className={cn(
                 "px-3 py-2 hover:bg-accent cursor-pointer flex items-center justify-between rounded-md",
-                isActive && "bg-accent",
+                activeFolder.toLowerCase() === folderName && "bg-accent",
               )}
-              onClick={() => router.push(`/dashboard/inbox/f/${path}`)}
+              onClick={() => router.push(`/inbox/f/${path}`)}
             >
               <div className="flex items-center space-x-3">
                 {folderName === "inbox" && (
