+++ apps/web/src/app/(authedLayout)/inbox/components/Tiptap.tsx
