+++ apps/web/src/app/(authedLayout)/inbox/f/[name]/page.tsx
