+++ apps/web/src/app/(authedLayout)/projects/page.tsx
