+++ apps/web/src/app/(authedLayout)/tasks/components/RedirectToTask.tsx
@@ -1,0 +1,3 @@
+import { redirect } from "next/navigation";
+
+export const RedirectToTasks = () => redirect("/tasks");
