--- apps/web/src/app/dashboard/page.tsx
@@ -1,7 +1,0 @@
-export default function App() {
-  return (
-    <article className="p-4 flex-1 h-full overflow-y-auto">
-      <p className="text-foreground">App page</p>
-    </article>
-  );
-}
