--- apps/web/src/app/dashboard/tasks/components/RedirectToTask.tsx
@@ -1,3 +1,0 @@
-import { redirect } from "next/navigation";
-
-export const RedirectToTasks = () => redirect("/dashboard/tasks");
